requirements.txt
# Python >= 3.8, < 3.12 
# Compatible with PyTorch ecosystem and modern features

torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
tensorboard>=2.13.0
torchmetrics>=0.11.0
segmentation-models-pytorch>=0.3.0
scikit-learn>=1.3.0 